LayoutTests/http/wpt/websockets/secure-cookie-secure-connection.https-expected.txt

Test WebSocket sends Secure cookies over secure connections.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

WebSocket open
WebSocket closed
PASS cookie is "WK-websocket-test-secure=1; WK-websocket-test=1"
PASS successfullyParsed is true

TEST COMPLETE
